Created BuzzerCode.ino, modified SensorReadout.ino & RGBCode.ino
BuzzerCode plays a noise through the buzzer. Unable to get the SensorReadout to work due to "Config error fxos8700" meaning the teensy and shield aren't connected properly. I will need to have better connecters and resources in the Invention Studio to further diagnose the issue. The RGB LED is also broken, but the function works.